Remove symbol O_SERVICE
It's sole purpose was to allow the `mkservice` command in the
src/cmd/ksh93/bltins/mkservice.c module to open a network endpoint in
listening mode. Since we removed that command the O_SERVICE symbol is no
longer needed.

I noticed this while working on PR #661 to eliminate Coverity Scan warnings
about reading past the end of the path name string. I decided to fix this
in a separate PR since PR #661 is already larger than I would like.