Add tests for special /dev/... path names
I intend to refactor `sh_open()` to a) fix lint warnings about accessing
bytes past the end of the path string and b) improve readability. Before
doing so it would be nice if there were tests of the existing behavior.

These tests are necessarily incomplete because we don't (yet) have a way to
force on or off the emulation code or pretend that paths like /dev/stdin
are not actually supported by the platform. This is a small step towards
creating a comprehensive set of tests for the magic /dev/... path names.