Fix reference past end of malloced buffer
ASAN brought this to our attention.

If var `name` points to a string like "schedule.job" then r == 8 (the
length of the first token). If the string pointed to by `nq->nvname` is
shorter than that then `nq->nvname[r]` will read a location past the end
of the buffer. This is yet another case where the person writing the
original version made a dangerous optimization.