Hack around another read past the end of heap buf
ASAN told us the buffer pointed to `sp` in the `memmove()` in `nv_putval()`
was reading past the end of the malloced buffer pointed to by `sp`. This
doesn't actually fix the bug. It merely papers over the problem documented
in issue #676. This is in the hope that by eliminating one problem other
problems may go away or be easier to debug.