Fix Invalid Lexing Of Parameter Expansions With Parens
This addresses #475 where ksh was found to incorrectly parse parameter
expansions with parentheses when the right-hand side of the expansion is a
string literal.

After a significant amount of analysis it was found that the lexer was mostly
functioning correctly, and that in this particular state it would actually do
the correct thing so long as it was placed in a mode where it would begin
treating the values following the parameter expansion characters as quoted.