Eliminate getenv() intercept
The libast intercepts of `getenv()` and `setenviron()` cause problems
for ASAN testing. Since we no longer support building and using libast
independent of ksh remove the indirection. This makes it possible to
enable ASAN on platforms like FreeBSD. Yes, this tightly couples libast
with ksh. But that is no longer a negative since the libast in this
project is not used outside of the `ksh` command.

Note that I could have, at least in theory, resolved the interaction
problems with ASAN without removing the intercept redirection which allows
libast to be used independent of the `ksh` program. But doing so seems
pointless since we no longer support building libast for any reason other
than building the `ksh` command.

I also make no claim that this is the minimal or optimal change to achieve
compatibility with tools like ASAN. But this does improve the usefulness of
ASAN on macOS and FreeBSD without introducing problems on Linux platforms.

Fixes #687