Fix more parameter expansion issues
Another attempting at closing #475.

After submitting pull-request #678, further issues were found with parameter
expansions involving the `:` character and the `+` character were found. This
appears to be due to the lexer being placed in the incorrect mode during
lexical skipping in the `varsub` method.

This fix now explicitly places the lexer in the correct mode when doing one of
the affected expansions. There may be a simpler or more elegant way to
accomplish this effect but it was not immediately obvious.