Weak symbols don't work on Cygwin
The `sh_getenv` weak symbol in libast does not satisfy references to it
on Cygwin. This appears to be a Cygwin bug. There are lots of complaints
about the behavior of weak symbols on Cygwin. Although I was unable to
find one that exactly matched the failure mode I observed. So rather
than depend on the weak symbol mechanism move the libast `sh_getenv()`
and `sh_setenviron()` functions into a separate `libenv` library and make
the normal "strong" symbols. Then link with that library when not building
ksh. This is the libast unit tests and the standalone `pty` command.