Small refactoring of sh_lex
The `sh_lex` function is currently a 1000+ line behemoth that severely needs to
be broken apart in order to reclaim any degree of intellectual understanding
and control.

This is a first, modest attempt at such a refactoring. Namely, the following
things are done to help clarify portions of the function:

* Add `S_NOP` lex state to indicate states that are not processed.
* Add `read_to_next_state_transition` function to encapsulate lexer behavior.