More cppcheck lint cleanups
The name.c module, specifically the nv_putval() function, should be
using the void pointer `vp` rather than the char pointer `sp` all the
places the pointer is to a data type other than a string. There are also
a lot of pointless parens and casts in those statements that can also be
removed.