Rename sh_lex String Variables
Rename the variable `state` in the `sh_lex` function to `state_table` to more
clearly reflect the fact that it is a pointer to a table rather than to a
particular state.

This change also adds a new variable `data` to take over in those situations
where `state` was doing double duty also holding a string of characters.