Remove `kill -Q n` flag
It makes zero sense for ksh to support including a value that looks like
a pointer when sending a queued signal via the `kill` command. This
is another case of the original team implementing a capability solely
because the underlying POSIX API supports it. Not because there was a
use case or it made sense in the context of a shell.

Resolves #709