Remove unused parameter from path_pwd
The `flags` parameter formerly passed to path_pwd was completely unused.
Therefore, it was a prime target for removal.

This change removes the `flags` parameter and updates all invocations of
`path_pwd` to no longer pass this value in. In addition, the documentation
above the function is updated to more accurately describe its current behavior.