Remove dead code in print.c
Coverity issue 253596 identified a section of dead code. The variable `neg` in
`extend` was assigned a constant `0` value and then unused and unassigned until
it was checked for a positive value in the removed `if` statement. Since the
constant value `0` would never evaluate to true, the section of code guarded by
the `if` would never execute.