Fix dead code in walk_tree
Fixes an issue originally injected in d9a8de2 where an unused variable `nq` led
to dead code. From what I can tell the variable was originally added to fix a
"free after use bug" (#398). Unfortunately the original declaration in a later
scope was not removed. This meant that rather than being assigned a value, the
more shallowly scoped `nq` was permanently hidden and never assigned a value,
as specified in C99 ISO/IEC 9899 6.2.1§4.

The fix prescribed here is to merely remove the later redeclaration of `nq`
allowing the shallow declaraction to both be assigned a value and successfully
freed if needed. Based on the above analysis this seems in line with the
original intent of the fix.