Tiny step to replacing AST locale subsystem
This is a very tiny step to replacing all the complexity of the AST
locale subsystem with the platform implementation. This effectively makes
`LC_OPTIONS=native` the default behavior.

This removes the wchar.sh unit test because it explicitly depends on the
`LC_OPTIONS` var which, while still technically valid with this change,
is not compatible with the goal of removing the AST locale subsystem.
Furthermore, the undocumented `printf "%(nounicodeliterals)q\n"` capability
is not present in the ksh93u+ release and I can see zero reason for
supporting it. The removal of that feature from the implementation will
be done in a subsequent commit.