Revert my earlier change to the order of the tests
It isn't enough that the `special-dev-paths` unit test is flagged to not be
run in parallel. Whether this is a Meson bug or expected behavior running
it in the middle of the other unit tests results in frequent failures. So
move it to the end of the list of unit tests as I had it originally.