Remove `_ast_codeset()`
Every system we're currently testing on or likely to run on supports
`nl_langinfo(CODESET)`. So just use that directly where `codeset()`
is currently used. Furthermore, that `codeset()` changes the platform
codeset string to a different string in some situations is dangerous. If
the mutated value is used in an `iconv_open()` it could fail.

Related #579