Remove feature symbol `_locale_utf8_str`
It turns out we can't rely on the value assigned to `_locale_utf8_str`
during the build config phase to match the value returned by
`nl_langinfo(CODESET)` at run time. So drop that feature test and
broaden the set of values that mean the current locale uses UTF-8.

Related #579