More removal of EBCDIC system support
This only eliminates the uses of CC_NATIVE which clearly are due to
determining if we're not building on an EBCDIC system. Removing that,
and related, symbols and header files will occur in a future change
because I need to explore if the AST iconv implementation can be
eliminated.

Related #742