Fix write after free error in history.c
This fixes Coverity #253710 in which we attempt to access the variable `hp`
after it has already been closed. This happens in the case where the predicate
`sfsync(hp->hisfp) < 0` evaluates to true. This results in a call to
`hist_close` which also frees `hp`.

Instead, we now avoid accessing `hp` altogether in the case where the predicate
is true, avoiding this issue.