Remove all support for building on EBCDIC systems
This removes the remaining code for building on platforms (e.g., IBM
mainframes) that use EBCDIC as its native encoding. All of that code
was turned into a no-op by the previous change so this simply removes
the headers and the modules which no longer provide any public APIs that
are used.

Resolves #742