`pathcat()` has nothing to do with 3DFS
While looking for places the code can be simplified I noticed the
comment in src/lib/libast/path/pathcat.c about how it is related to 3D
FS support. It turns out that was true of code in that module in the
past. It is no longer true. So remove the misleading comments.