Fix issue found by clang static analysis
Oclint reported

src/cmd/ksh93/bltins/cd_pwd.c:228:9: Value stored to 'rval' is never read

Which is true for platforms that do not implement the O_SEARCH symbol.
Such as macOS and Linux.  This reorgs the code slightly to make that
assignment conditional on the platform implementing O_SEARCH.