Fix serious lint about an invalid use of NULL
Cppcheck warned us:

[src/cmd/ksh93/sh/parse.c:1381] warning (nullPointerArithmeticRedundantCheck):
 Either the condition 'last' is redundant or there is overflow in pointer subtraction.

That is because the block of code in question appears to do the equivalent
of `NULL - ptr` under some conditions to calculate the number of chars
between two pointers. In reality the initial `last = strchr(...);`
statement has to always find the desired char or there is an API
violation. So rewrite the code to make the clearer.