Remove `streq()`
Remove unneeded and possibly counterproductive `streq()`. The hand
optimization it performs may or may not be a net win. And the clearer
meaning compared to `strcmp()` is nice it also means anyone looking at the
code will have to search the code to realize it is just a wrapper around
`strcmp()`.