Remove `alarm` builtin
The `alarm` builtin has been present and undocumented for at least 10
years. Dating back to the ksh93s release if not earlier since the 93s
beta release appears to have been made on 2006-12-29. Since it is old,
undocumented, and a public statement by the AST team suggests it is not
ready for prime time it is time to remove it in the interest of simplifying
the project.

Resolves #646