Reinstate two `newof()` uses
My previous changes made two changes to `newof()` invocations with a non
constant NULL for the first argument. That is dangerous. So even though
all unit tests are passsing revert those two changes.

P.S. Note that the realloc behavior is itself problematic. That's because
the AST Vmalloc subsystem would zero initialize the new bytes but no
platform (GNU, BSD, or other) that I'm familiar with will do so.