Remove var `viraw`
While working on lint cleanups I noticed that viraw is now a constant
thanks to oclint warning about constant if() expressions. We no longer
support the vi CLI editor mode without support for putting the tty in
"raw" mode. So eliminate the symbol. And update the documentation to
explain that the option no longer has any effect.

Fixes #864