Fix Coverity false positive
Coverity warned that theoretically the result of calling `nv_getn()`
could be negative. And using that negative value to index an array is
dangerous and certainly wrong. So add an `assert()` to silence the
warning and ensure that if a negative value was about to be used it
kills the shell.

CID#253818