Don't return a pointer to a stack var
Coverity Scan pointed out that it is possible for `base64encode()` to
assign the address of `tmp` to the `tn` argument. Since no users of this
function actually use the `tn` parameter just remove it.

This also fixes a related failure mode that clang static analysis
identified after removing `tn`:

src/lib/libast/string/base64.c:124:51: The left operand of '==' is a
garbage value

So predicate that test on `tp != tmp`.

This also enables running the API/base64.c unit test.

CID#253637
Fixes #890