Lint: SFOPEN() definition is confusing
The SFOPEN() macro causes lots of lint warnings like this one:

src/lib/libast/sfio/sfstack.c:104:5: constant conditional operator [basic|P2]

That's because the majority of uses pass a constant zero for the second,
`l`, parameter which turns the macro into a no-op. It also obscures the
purpose of the second parameter. Redefine the macro to take just a
single parameter and redefine the few places which don't want it to
always open a SFIO stream by turning

    SFOPEN(f, local)

into

    if (!local) SFOPEN(f);

This both eliminates the warnings and clarifies the behavior.

Fixes #898