Coverity: mem leak
This should eliminate a memory leak identified by Coverity. The changes
of `calloc()` to `malloc()` are because the memory doesn't need to be
zeroed. That's a vestage from when the code used `newof()` which we
replaced with `calloc()` to retain the newof semantics. Surprise! Some
of those `newof()` uses didn't need to clear the memory.

CID#288384
Fixes #902