More lint from SFIO macros
The problem is that too many of these macros are written in ways that cause
static analysis to complain about things like unreachable empty statements.

Simply converting all of them to the `do { ... } while(0)` pattern
resolves a lot of the lint warnings. However doing that to one of the most
frequently used macros, `SFMTXRETURN`, introduces a different lint warning
about a branch statement as the last statement of a loop. So instead of
treating these macros as function calls which need to be followed by a
semicolon just drop the semicolon where they macro is used. After all, most
of them cannot be turned into functions precisely because they do things,
such as `return`, that have to be done in the context where they are used.

This eliminates 100 lines of warnings from `bin/lint all`.

P.S., I consider `return` statements within macro definitions an
anti-pattern; i.e., bad practice. But that is something I'm not going to
change now as this is already quite invasive.