Replace unneeded `ASSERT()` macro
There are just a handful of places that use the `ASSESRT()` macro
defined in sfhdr.h. Replace it with the standard `assert()` macro.

Also, remove the silly `#define SECOND 1000` macro both because the
symbol name doesn't tell the reader anything useful about the value and
because the handful of places it is used simply using the magic `1000`
is perfectly clear.