Eliminate some static analysis lint
The issues here are likely false positives. Such as this one:

```
src/lib/libast/misc/stk.c:274:32: Access to field 'stkoverflow' results
in a dereference of a null pointer (loaded from variable 'sp')
```

But restructure the code to suppress the warning and make it more
obvious what the constraints are.