Remove lint from module io.c
Clang static analysis reports:

src/cmd/ksh93/sh/io.c:541:14: Potential leak of memory pointed to by 's'

This eliminates that lint warning and also restructures the code to clarify
the behavior. As always, this type of change is not meant to address
every shortcoming of the code being modified. It's only meant to be an
incremental step to improving the quality by eliminating a lint warning.

Related #853