Remove LIBICONV_PLUG
There was a period, roughly a year ago, when we needed to define
preprocessor symbol LIBICONV_PLUG on some platforms to build ksh. That
causes problems when building ksh via MacPorts. And testing on the
same platforms I was using a year ago when I introduced this hack shows
that it is apparently no longer needed. Presumably the real problem lay
elsewhere. Most likely in the various AST wrappers and `#define` symbols
that we've removed in the intervening time.

Resolves #932