Lint cleanup
src/lib/libast/regex/regstat.c:44:25: Access to field 'type' results in
a dereference of a null pointer (loaded from field 'rex')

The `p && p->env` conditions don't make any sense since both have to be
non-NULL at that point. However, it is possible for `p->env->rex` to be
NULL so that needs to be tested before dereferencing it to get the
`type` value.