Lint cleanup
src/lib/libast/regex/regnexec.c:1798:23: The right operand of '>' is a
garbage value

It's theoretically possible for the `i > n` condition to be evaluated
without `n` being assigned to. It probably can't happen in practice but
it's easy enough to initialize `n` where it is defined to ensure we
don't use whatever random value was on the stack.