Replace stpcpy() with strlcpy()
This only does so for one place where it matters. It replaces `stcpy()`
with `strcpy()` where the return value doesn't matter. It leaves the uses
in getopt.c because we want to eliminate that code. It leaves the other
uses in glob.c because a simple replacement of one function for another
doesn't address the core problems with those uses of stpcpy.

Resolves #956