Fix assumption that chars are unsigned
The code in question incorrectly assumes either that a) characters in
the string being scanned will never include a codepoint greater than 127
and/or b) that the char type is unsigned. Fix that invalid assumption.

This does not purport to fix all such bugs in the lexer. Only the specific
failure documented in GitHub issue #971.

Fixes #971