Require a compatible version of clang-format
Another contributor, @siteshwar at RedHat, recently implemented some
automation to run `bin/style --all` on a regular basis. Unfortunately the
system he ran it on has an old version of `clang-format` which produces
incorrect formatting. Or at least inconsistent with newer versions of
`clang-format` which causes unwanted thrashing of the code style. So
impose a minimum acceptable version of the `clang-format` tool.