rename regex.h and simplify the header
After upgrading my main macOS server to 10.14, Mojave, along with clang,
iwyu, etc. I'm now seeing lots of iwyu warnings due to the conflict
between the AST and system regex.h headers. So rename the AST header and
remove the symbols that regulate the order of the includes as they're no
longer needed.