Fix previous Coverity fix
Commit 993fc6aa was the wrong fix for Coverity CID#279500. The real
problem was not the structure length it was passing a pointer to the
first member of the structure rather than a pointer to the structure.
Yes, the two are the same but Coverity correctly points out that if you
really only want the structure member to be copied then the length is
wrong.