Remove all support for fixed size arrays
Commit 54cd98f2 removed support for fixed size arrays by removing the
`SHOPT_FIXEDARRAY` symbol and code predicated on it. While looking into
how to resolve issue #1002 I noticed there is a structure member and a
few bits of code that are no longer relevant as a result of that earlier
change. This change completes the work started by that previous commit.

Related #1002