Remove namval.h
The `Namval_t` structure defined in src/lib/libast/include/namval.h has
absolutely nothing to do with the structure of the same name that is the
core structure for ksh variable handling. This unrelated structure is
used in just three modules to define lists of options for the `stropt()`
function.  There isn't any reason to maintain a separate header for
this trivial structure. So define it in each of the three modules that
use it and give it a different name to avoid confusion with the other,
far more important, `Namval_t` type.