Cleanup some NV_* symbols
While investigating the use of NV_JSON in the code I noticed that the
definition of `NV_UINT16P` is broken. Which has never been noticed because
it is unused.

But that also reminded me that all the namval attribute bit definitions
(e.g., NV_RDONLY and NV_NOFREE) are defined in terms of decimal or hex
constants and are in a random order. Which makes it hard to recognize
which bits are still unused. So change those definitions in terms of bit
shifts and order them from lsb to msb; e.g., 0x200 becomes 1 << 9.

This does not change the binary definition of any values; only their
symbolic representation in the source. This is just the first of many
changes in this vein I want to make.  Mostly because given how the code is
written it is extremely dangerous and confusing to overload the definition
of specific bits. For example, aliasing `NV_SHORT` to `NV_RJUST` and
relying on whether `NV_INTEGER` is set to disambiguate the two cases.