Remove `_NV_PRIVATE`
The `_NV_PRIVATE` symbol is not needed. It only serves to mask circular
imports, make understanding the code harder, and introduce potential bugs.

If and when this project again makes it possible to build plugins without
using the entire source tree it is unlikely we will want to have `struct
Namval` members that are not visible to the plugin. At least not without
using a pointer to an opaque structure. But if we do we can do so in a
manner that is less problematic.