Rename `NV_IDENT` to `NV_MISC`
The code incorrectly overloads `NV_IDENT`. This is a tiny step to
separating some of those uses. Primarily by making it clear that `NV_IDENT`
is a nvflag bit value by aliasing it to the newly introduced `NV_MISC`
symbol. To be done is actually separating the two uses since there is no
reason they should be coupeled in this fashion.

Related #1038