Remove NV_PUBLIC
The definition of `NV_PUBLIC` is wrong. It includes two symbols,
`NV_NOSCOPE` and `NV_NOADD`, that can never appear in the `nvflag` struct
member. So it doesn't make any sense to include them given how `NV_PUBLIC`
is used.

It is a no-op for two of the three places it is used. For the third use
all it does is filter out `NV_NOFREE` from the flag so do that explicitly.

Looking at the ksh93u+ branch its use appears to be broken there as
well. So the code has been this way for quite a while.

Related #1038