Introduce nv_associative() operation type
This introduces a Nvassoc_op_t type and renames the related symbols to
avoid any confusion with the `NV_*` symbols used that define the attributes
of a name/value object. It makes the type a struct rather than a simple
int to ensure that any attempt to pass a simple int will be caught by
the compiler.

Related #1063