`bin/style all` plus bump to clang-format version
I just ran  `bin/style all` on my development server and was surprised
by the large number of changes. That's because my system now has
clang-format 8.0.0 which finally stops producing output like this for
`switch` blocks just because the statement fits on a single line:

    default: { break; }

It now, more sensibly, formats that as

    default: {
        break;
    }

So commit those style changes and bump the minimum acceptable clang
version to 8.0.0.