Replace HASH_NOSCOPE with NV_NOSCOPE
The HASH_NOSCOPE symbol is a holdover from the pre-CDT hash library.  Given
that it is combined with other name/value symbols like NV_ADD it should
be put in that namespace. In fact, there is already an equivalent symbol
in that namespace: NV_NOSCOPE. So replace HASH_NOSCOPE with NV_NOSCOPE.
This removes a potential source of bugs and makes the logic clearer.

Related #1072