Eliminate symbol HASH_BUCKET
The HASH_BUCKET symbol is a holdover from the hash library that predates
the CDT subsystem. It's sole purpose it to communicate to nv_search()
that its first argument is a Namval_t* rather than a char*. Using a flag
and explicit casts to do that is dangerous, confusing, and tends to make
it harder for the compiler and lint tools to warn about potential bugs.

So refactor nv_search() into two distinct interfaces. One that takes a
Namval_t* and one that takes a char*. Thus eliminating the need for the
symbol HASH_BUCKET.

This would have been a smaller change but I deliberately chose to make
some of the function parameters "const" to help the compiler and lint
tools warn about potential problems. Something we should do more of.

Resolves #1072