Restructure the `basic.sh` unit test
It turns out that none of the tests predicated on a bare `cat` invoking
the builtin depend on that exact assertion. Too, the only test within that
block which needs such an assertion does not use `cat` and is instead
really predicated on whether or not the system supports the `/dev/fd/`
directory.

Related #951