Replace ERRIO with STDERR_FILENO
While working on fixing issue #1079 I noticed the ERRIO symbol that is
simply an obscure alias for the standard STDERR_FILENO symbol. So replace
the former with the latter.

If we ever get a build failure on a system that does not provide the
industry standard STDERR_FILENO symbol we can implement a build time test
for its availability and provide a fallback. Although in that case we
should also ensure the same magic constant via a compile time symbol is
also used in the SFIO code. Hint: At the moment ERRIO is unique to the
ksh code (i.e., not used in the SFIO code) and is not defined in terms
of a SFIO constant.

Related #1079