More nv_open() safety + silence a Coverity warning
There is a theoretical path from extend() (in print.c) thru fmtbase64()
to nv_open() that can result in the latter accessing a byte past the end
of an empty string. It doesn't look like that can actually happen due
to constraints that Coverity isn't aware of.  There are, however, ways
(most notably from the nv_open() in put_tree()) that a zero length var
name can be passed and which expect a NULL value to be returned
(verified by instrumenting the code).

Investigating this caused me to notice that nv_open() only handles a zero
length var name more or less by accident. Mostly by virtue of nv_search()
returning a NULL pointer for that case. Instead explicitly test for a
nil var name and return early. Hopefully this will silence the Coverity
warning.

Coverity CID#253849