Fix ast_temp_path() implementation
When I refactored ast_temp_file() to eliminate the use of mktemp()
I implemented ast_temp_path(). But the assignment of the path to
`shp->fifo` lead me to believe it was sufficient to generate a path
that was unique to the shell. That assumption was wrong. A unique path
needs to be generated each time it is called. This fixes several tests
in src/cmd/ksh93/tests/basic.sh involving multiple process substitutions
that fail on platforms, like FreeBSD, which don't support /dev/fd.