Refactor sfpkrd() to use poll() or select()
There is no justification for falling back to `select()` if the `poll()`
call fails with an unexpected errno value. And testing shows that never
happens on any of the many platforms I use to test changes to the project.
So refactor the code to use one or the other but not both.

Add a test for the situation which resulted in issue #1091 being opened.

Fixes #1091