Remove bogus code
The code in sfpkrd() to detect if we're on macOS 10.4 is broken. Calling
recv() on a regular pipe always fails with errno == ENOTSOCK. Including
on other platforms like Linux. Not to mention macOS 10.4 is ancient
history that will never see a new version of ksh. So remove that broken,
dead, code.

Related #1091