Another Coverity bug fix
Coverity warns that the fmtre() function might be called with a string
that is less than three chars in length which it will then dereference
past the end of the string. Not likely to be an issue in practice but
the too short string could theoretically end on a page boundary with the
following page being invalid. Which will cause a SIGSEGV.

Coverity CID#253849