Change some API test macros to avoid lint
Recent changes that converted preprocessor macros to static inline
functions causes lint tools like cppcheck to issue warnings like this:

[src/lib/libast/include/sfio.h:322] warning (nullPointerRedundantCheck):
 Either the condition '!f' is redundant or there is possible null pointer dereference: f.

That is actually from src/cmd/tests/sfio/tnotify.c because of how
`terror()` is defined. Redefine that, and related, macros into a form
that that allows cppcheck to recognize the `__attribute__((noreturn))`
of the actual function that is run.