Tweak the definition of `assert()`
Recent changes to several SFIO symbols that used to be preprocessor
macros and are now static inline functions causes cppcheck to warn about
"possible null pointer dereference". That's because it doesn't recognize
that an `assert()` ensures that dereference can't happen. This changes
the definition of `assert()` to silence those warnings.