Eliminate tests for external commands
We no longer build external commands like `fmt`. To the extent that we
still build a few commands as ksh builtins (e.g., `cat`) we will need
to write new unit tests from scratch. Whomever does so may look to the
legacy tests for inspiration. But there is no good reason for these files
to remain in the source tree and pollute the output of greping the source.