Lint: iousepipe() return value is a constant
Both clang static analysis and Coverity Scan reported that the value
assigned to `unpipe` for the `if (shp->subshell) {` case is unused when
building when `SHOPT_SPAWN` is false. There are only two callers of
`iousepipe()` and one of them ignores the return value which makes sense
since the return value is a constant -- once you realize that `sh_rpipe()`
can't fail. For the call site that does use the return value just set the
`unpipe` to true only if `SHOPT_SPAWN` is true to mirror where the value of
`unpipe` is used.

Note that while `sh_rpipe()` can fail it cannot return a failure value.
For good or bad it always returns success unless there is a problem in
which case it longjmp's via the call to `errormsg()`.