Add run_lsof() debug function
This does two things: a) add the current pid to debug messages, and b)
add a run_lsof() function to do just that. I've found myself needing
the latter many times in the past year to help figure out what is going
on vis-a-vis open file descriptors. And working on issue #1117 made me
decide to formally add the code.

The addr to file+line cache is introduced but not used by this change.