Remove Solaris specific hack
At the moment we can't even build on Solaris since it doesn't support
Meson. But even if we could it is dubious that we should be using a
Solaris specific feature to modify the increments in which the stack
grows. Especially now that we're no longer using the AST vmalloc
subsystem and are instead using the platform malloc subsystem.