Make use of posix_spawn configurable
Changes since we disabled the use of posix_spawn() by default have resulted
in bit rot. Make use of posix_spawn() configurable at build time and fix
some of the code rot.

Note that this produces a working ksh but the builtin cat causes a SIGSEGV
when the `sfeof()` on line 479 of src/lib/libcmd/cat.c is executed. TBD is
what is wrong but at least now it's easy to build ksh with posix_spawn()
enabled so we can find and fix such problems.