Remove definitions of F_DUPFD_CLOEXEC
It's defined in two incompatible ways in two headers whose order of
inclusion is undefined. Too, these fallback definitions are not needed
on any platform we're currently validating that things work; including
CYGWIN. If we find a platform where a fallback is needed we can revisit
doing this in a safer manner.