Remove stropt()
The stropt() implementation is a large block of code that has too much
lint and what appears to be at least one bug. It's only used three
places -- none of which are documented. Only the use in the error.c
module might have been worth keeping and I'd rather replace that use
with a more straightforward mechanism for debugging messages.

Resolves #937