Remove _SH_PRIVATE and fix related bogosities
There is simply no way in C to make structure members truly private.
Especially in light of the fact the source code for the project is
public. Private code complicates the code and requires that users know
they have to include headers in a specific order rather than just
alphabetically (or a random order). This is the first step in fixing
that situation.

Removing _SH_PRIVATE resulted in several other problems being found. For
example, these declarations in shell.h:

    extern void sh_offoption(Shell_t *, int);
    extern bool sh_isoption(Shell_t *, int);
    extern void sh_onoption(Shell_t *, int);

Yet those are not external functions. They are, in fact, macros defined
in defs.h.