Fix unit test that kills unexpected processes
While working on my previous commit I was surprised to see my editor
(whose command line included the file `sleep.c`) was being killed when
running unit tests. The reason is that one of the tests in the functions.sh
unit test kills any process with the word "sleep" in its command line.

This change also shaves a couple of seconds off that unit test.