Remove _typ_long_double feature symbol
We haven't been testing if symbol `_typ_long_double` should be defined.
Yet on any platform we currently care about symbol _typ_long_double
would be defined to be "1" (i.e., true). So eliminate it and
unconditionally compile the one block of code, a unit test, that depends
on it.