Restructure code to eliminate cppcheck warning
The existing code was written in a manner which confuses both human readers
and lint tools like `cppcheck` due to `mb1char()` being a preprocessor
macro with unexpected side-effects. This change only fixes the issue for
tools like `cppcheck`. For humans I've already noted in a previous change
that use of `mb1char()` needs to be replaced. See the ast.h header.

Fixes #1138