Remove dangerous #define statements in sfhdr.h
The block of defines starting with

    #define _next next

are extremely dangerous as they affect any use of those symbols; not
just those in struct _sfio_s. They also needlessly obfuscate the code.
Not to mention that the underscore symbols aren't really private. There are
lots of direct uses in the ksh and other, non-sfio implementation, code.
Which make the distinction between the underscore and non-underscore names
meaningless. Too, those defines mean that if you happen to include sfio.h
before sfhdr.h the code fails to compile and thus makes the code fragile
for that reason as well.

So remove those definitions and just use the non-underscore variant
everywhere.