Change `union Value` into a struct
This changes `union Value` into a struct with getter/setter functions.
It only fixes a couple of blatant type mismatches I found while replacing
the direct member access with a getter/setter function. Such as the couple
I documented in issue #1150.

It also fixes one blatant bug in the calculation of the size of a
dynamically allocated buffer to hold instances of the union/struct that
could cause the allocation to be too small. The bug was in `array_grow()`
using `sizeof(union Value *)` when it should have been `sizeof(union Value)`.

Related #1150